:sparkles: Feature: reorganize web/vscode structure to extract pages definition
