+++ src/extension.ts
@@ -3,7 +3,7 @@ import * as fs from 'fs'
 import * as path from 'path'
 import * as vscode from 'vscode'
 import VSPicgo from './vs-picgo'
-import { PanelManager } from './webview/PanelManager'
+import { PanelManager } from './vs-picgo/PanelManager'
 
 async function uploadImageFromClipboard(vspicgo: VSPicgo) {
   return await vspicgo.upload()
@@ -62,6 +62,9 @@ export async function activate(context: vscode.ExtensionContext) {
 
     vscode.commands.registerCommand('picgo.webviewDemo', () =>
       panelManager.createOrShowWebviewPanel('Demo')
+    ),
+    vscode.commands.registerCommand('picgo.webviewPicGoControlPanel', () =>
+      panelManager.createOrShowWebviewPanel('PicGoControlPanel')
     )
   ]
   context.subscriptions.push(...disposable)
