+++ src/utils/page.ts
@@ -1,0 +1,5 @@
+export const pageMap = {
+  Demo: 'Demo',
+  PicGoControlPanel: 'PicGo Control Panel'
+}
+export type PageId = keyof typeof pageMap
