+++ src/vs-picgo/template.html
