+++ src/webview/pages/PicGoControlPanel.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+import './PicGoControlPanel.less'
+
+export const PicGoControlPanel = () => {
+  return <div className="picgo-control-panel">PicGo</div>
+}
