+++ src/webview/pages/index.tsx
@@ -1,15 +1,21 @@
+/**
+ * This script will be loaded by the template html used in webview, see PanelManager.getPageHtml
+ */
 import React from 'react'
 import ReactDOM from 'react-dom'
 import { showMessage } from './channel'
 import { Demo } from './Demo'
+import { PicGoControlPanel } from './PicGoControlPanel'
+import { PageId } from '../../utils/page'
 
 import './index.less'
 
 const pages = {
-  Demo
+  Demo,
+  PicGoControlPanel
 }
 
-export const renderApp = (pageId: keyof typeof pages) => {
+export const renderApp = (pageId: PageId) => {
   const root = document.getElementById('root')
   if (!root) {
     showMessage({
