+++ src/webview/tsconfig.json
@@ -4,6 +4,5 @@
     "lib": ["DOM", "DOM.Iterable", "ESNext"],
     "noEmit": true,
     "jsx": "react"
-  },
-  "include": ["**/*"]
+  }
 }
